requirements.txt
typing-extensions
selenium
python-dotenv
# Add other dependencies if needed

